Move Bundle namespace classes into Bridge\Symfony
