+++ UPGRADE-2.1.md
@@ -1,0 +1,7 @@
+# Upgrade from 2.0 to 2.1
+
+## Deprecated Bundle namespace
+
+The `SLLH\IsoCodesValidator\Bundle` namespace is deprecated.
+
+Use `SLLH\IsoCodesValidator\Bridge\Symfony` namespace instead.
