+++ phpunit.xml.dist
@@ -18,7 +18,7 @@
         <testsuite name="symfony">
             <directory>./tests/Constraints</directory>
             <file>./tests/AbstractIsoCodeGenericConstraintTest.php</file>
-            <directory>./tests/Bundle</directory>
+            <directory>./tests/Bridge/Symfony</directory>
         </testsuite>
         <testsuite name="silex">
             <directory>./tests/Constraints</directory>
