+++ src/Bridge/Symfony/Bundle/SLLHIsoCodesValidatorBundle.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace SLLH\IsoCodesValidator\Bridge\Symfony\Bundle;
+
+use SLLH\IsoCodesValidator\Bridge\Symfony\DependencyInjection\Compiler\TranslationPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+class SLLHIsoCodesValidatorBundle extends Bundle
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new TranslationPass());
+    }
+}
