+++ src/Bundle/SLLHIsoCodesValidatorBundle.php
@@ -2,22 +2,17 @@
 
 namespace SLLH\IsoCodesValidator\Bundle;
 
-use SLLH\IsoCodesValidator\Bundle\DependencyInjection\Compiler\TranslationPass;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\HttpKernel\Bundle\Bundle;
+@trigger_error(
+    'The '.__NAMESPACE__.'\SLLHIsoCodesValidatorBundle class is deprecated since version 3.1 and will be removed in 4.0.'
+    .' Use SLLH\IsoCodesValidator\Bridge\Symfony\Bundle\SLLHIsoCodesValidatorBundle instead.',
+    E_USER_DEPRECATED
+);
 
 /**
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ *
+ * @deprecated since version 2.1, to be removed in 3.0. Use SLLH\IsoCodesValidator\Bridge\Symfony\Bundle\SLLHIsoCodesValidatorBundle instead.
  */
-class SLLHIsoCodesValidatorBundle extends Bundle
+class SLLHIsoCodesValidatorBundle extends \SLLH\IsoCodesValidator\Bridge\Symfony\Bundle\SLLHIsoCodesValidatorBundle
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function build(ContainerBuilder $container)
-    {
-        parent::build($container);
-
-        $container->addCompilerPass(new TranslationPass());
-    }
 }
